refactor(frontend): Move markers from Scenario- to AutothemeCodeMirror
