+++ frontend/src/app/shared/autotheme-codemirror/autotheme-codemirror.component.ts
@@ -3,6 +3,7 @@ import {CodemirrorComponent} from '@ctrl/ngx-codemirror';
 import {ThemeService} from 'ng-bootstrap-darkmode';
 import {of, Subscription} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
+import {Marker} from '../model/marker';
 
 @Component({
   selector: 'app-autotheme-codemirror',
@@ -15,6 +16,8 @@ export class AutothemeCodemirrorComponent implements OnInit, OnDestroy {
 
   @Input() options: any;
 
+  private _markers: Marker[] = [];
+
   @ViewChild('ngxCodemirror') ngxCodemirror: CodemirrorComponent;
 
   private subscription: Subscription;
@@ -22,6 +25,24 @@ export class AutothemeCodemirrorComponent implements OnInit, OnDestroy {
   constructor(
     private themeService: ThemeService,
   ) {
+  }
+
+  get markers(): Marker[] {
+    return this._markers;
+  }
+
+  @Input()
+  set markers(value: Marker[]) {
+    this._markers = value;
+    this.options.lint ??= {
+      lintOnChange: false,
+      getAnnotations: () => this._markers,
+    };
+    this.performLint();
+  }
+
+  private performLint() {
+    (this.ngxCodemirror?.codeMirror as any)?.performLint?.();
   }
 
   ngOnInit() {
