+++ frontend/src/app/shared/scenario-codemirror/scenario-codemirror.component.html
@@ -3,5 +3,6 @@
   [content]="content"
   (contentChange)="content=$event; contentChange.emit($event)"
   [options]="options"
+  [markers]="markers"
 >
 </app-autotheme-codemirror>
